requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
tavily-python>=0.3.0
edge-tts>=6.1.0
pydub>=0.25.1
audioop-lts>=0.2.1; python_version >= "3.13"
